Push can_edit and can_delete down from BaseComment to SectionComment
At the time of the commit, these are specific to SectionComment's context. BaseComment doesn't even have a parent set, so it would probably crash.

Can later add them as abstract methods to BaseComment if needs be.
