+++ democracy/models/comment.py
@@ -102,25 +102,6 @@ class BaseComment(BaseModel):
         if self.parent_id:  # pragma: no branch
             self.parent.recache_n_comments()
 
-    def can_edit(self, request):
-        """
-        Whether the given request (HTTP or DRF) is allowed to edit this Comment.
-        """
-        if request.user.is_authenticated and self.created_by == request.user:
-            # also make sure the hearing is still commentable
-            try:
-                self.parent.check_commenting(request)
-            except ValidationError:
-                return False
-            return True
-        return False
-
-    def can_delete(self, request):
-        """
-        Whether the given request (HTTP or DRF) is allowed to delete this Comment.
-        """
-        return self.can_edit(request)
-
 
 def comment_recache(sender, instance, using, created, **kwargs):
     """
